move `VoigtNotation` and `FullySymmetric4Tensor` to `sofa::type`, update namespaces and references accordingly
